+++ src/main/java/ru/yandex/practicum/filmorate/annotations/IsCorrectLocalData.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 import static java.lang.annotation.ElementType.FIELD;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
-@Target({ FIELD })
+@Target({FIELD})
 @Retention(RUNTIME)
 @Constraint(validatedBy = IsCorrectReleaseDate.class)
 @Documented
@@ -20,10 +20,14 @@ public @interface IsCorrectLocalData {
 
     String message() default "{IsCorrectLocalData.invalid}";
 
-    Class<?>[] groups() default { };
+    Class<?>[] groups() default {};
 
-    Class<? extends Payload>[] payload() default { };
+    Class<? extends Payload>[] payload() default {};
 
-    String date();
+    int year();
+
+    int month();
+
+    int day();
 
 }
