+++ src/main/java/ru/yandex/practicum/filmorate/controllers/UserController.java
@@ -16,21 +16,22 @@ import java.util.List;
 @Validated
 @RestController
 @Slf4j
+@RequestMapping("/users")
 public class UserController {
     private final HashMap<Integer, User> userStorage = new HashMap<>();
     private Integer counter = 1;
 
-    @GetMapping("/users")
+    @GetMapping
     public List<User> getAllUsers() {
         log.info("GET enabled. List of users was successfully sent.");
         return new ArrayList<>(userStorage.values());
     }
 
-    @PutMapping("/users")
-    public User putUser(@Valid @RequestBody  User user) {
+    @PutMapping
+    public User putUser(@Valid @RequestBody User user) {
         if (!userStorage.containsKey(user.getId())) {
             log.error("Cannot find user with this id. User with id " + user.getId() + " was not be replaced.");
-            throw new BeanNotFoundException("User with id="+user.getId()+" not found");
+            throw new BeanNotFoundException("User with id=" + user.getId() + " not found");
         } else {
             log.info("Used POST-method. User with id " + user.getId() + " was added.");
             userStorage.put(user.getId(), user);
@@ -38,20 +39,20 @@ public class UserController {
         }
     }
 
-    @PostMapping("/users")
+    @PostMapping
     public User setUser(@RequestBody @Valid User user) {
         if (userStorage.containsKey(user.getId())) {
             log.error("User with this id is already added. User was not be added.");
-            throw new BeanAlreadyCreatedException("User with id="+user.getId()+" is already added.");
+            throw new BeanAlreadyCreatedException("User with id=" + user.getId() + " is already added.");
         } else {
-            if (user.getId() == null){
+            if (user.getId() == null) {
                 user.setId(generateId());
             } else {
-                if (user.getId()>counter){
-                    counter=user.getId();
+                if (user.getId() > counter) {
+                    counter = user.getId();
                 }
             }
-            if (user.getName().equals("")){//поставил @NotNull в поле name у User, а если логин=null - то до сюда и не
+            if (user.getName().equals("")) {//поставил @NotNull в поле name у User, а если логин=null - то до сюда и не
                 user.setName(user.getLogin());                                                //  дойдёт(NotValid)
             }
             log.info("Used POST-method. User with id " + user.getId() + " was added.");
@@ -60,7 +61,7 @@ public class UserController {
         }
     }
 
-    private Integer generateId(){
+    private Integer generateId() {
         return counter++;
     }
