+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -6,6 +6,7 @@ import ru.yandex.practicum.filmorate.annotations.IsCorrectLocalData;
 
 import javax.validation.constraints.*;
 import java.time.LocalDate;
+
 @Data
 @AllArgsConstructor
 public class Film {
@@ -13,9 +14,10 @@ public class Film {
 
     @NotBlank(message = "Имя не должно быть пустым")
     private String name;
-    @Size(max = 200,message = "Описание не должно превышать 200 символов.")
+    @Size(max = 200, message = "Описание не должно превышать 200 символов.")
     private String description;
-    @IsCorrectLocalData(date = "1895-12-28",message = "Дата релиза не должна быть ранее 28 декабря 1895 года.")
+    @IsCorrectLocalData(year = 1895, month = 12, day = 28,
+            message = "Дата релиза не должна быть ранее 28 декабря 1895 года.")
     private LocalDate releaseDate;
     @Positive(message = "Длительность фильма должна быть положительной.")
     private Long duration;
