+++ src/test/java/ru/yandex/practicum/filmorate/FilmControllerTests.java
@@ -4,11 +4,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.validation.annotation.Validated;
 import ru.yandex.practicum.filmorate.controllers.FilmController;
-import ru.yandex.practicum.filmorate.controllers.UserController;
 import ru.yandex.practicum.filmorate.exceptions.BeanAlreadyCreatedException;
 import ru.yandex.practicum.filmorate.exceptions.BeanNotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.model.User;
 
 import javax.validation.*;
 import java.time.LocalDate;
@@ -19,7 +17,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @SpringBootTest
 @Validated
-class FilmorateControllersTests {
+class FilmControllerTests {
 
     @Test
     void getFilmsTest() {
@@ -42,11 +40,7 @@ class FilmorateControllersTests {
         assertEquals(filmDurationException.getMessage().substring(10),
                 "Длительность фильма должна быть положительной.");
         notCorrectFilmAquaMan.setDuration(133L);//возвращаем положительное значение длительности
-        notCorrectFilmAquaMan.setDescription("qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqq" +
-                "qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqq" +
-                "qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqq" +
-                "qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqq" +
-                "qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqq");
+        notCorrectFilmAquaMan.setDescription("q".repeat(250));
         final ConstraintViolationException filmDescriptionException = assertThrows(
                 ConstraintViolationException.class, () -> validateInput(notCorrectFilmAquaMan));
         assertEquals(filmDescriptionException.getMessage().substring(13),
@@ -85,83 +79,10 @@ class FilmorateControllersTests {
         fc.putFilm(aquaMan2);
         assertEquals(fc.getAllFilms().get(0), aquaMan2);
     }
-
-    @Test
-    void getUsersTest() {
-        UserController uc = new UserController();
-        User user = new User(1, "JavaGod", "Hacker", "java@yandex.ru",
-                LocalDate.of(1990, 12, 12));
-        uc.setUser(user);
-        assertEquals(uc.getAllUsers().get(0), user);
-    }
-
-    @Test
-    void setUsersTest() {
-        UserController uc = new UserController();
-        User notCorrectUser = new User(1, "JavaGod", "Hacker","java@yandex.ru",
-                LocalDate.of(1990, 12, 12));
-        notCorrectUser.setLogin("");
-        final ConstraintViolationException loginEmptyException = assertThrows(
-                ConstraintViolationException.class, () -> validateInput(notCorrectUser)); //exception message: "login: Логин не должен быть пустым"
-        assertEquals(loginEmptyException.getMessage().substring(7),
-                "Логин не должен быть пустым");
-        notCorrectUser.setLogin("Java God");
-        final ConstraintViolationException loginWithSpasesException = assertThrows(
-                ConstraintViolationException.class, () -> validateInput(notCorrectUser));
-        assertEquals(loginWithSpasesException.getMessage().substring(7),
-                "Логин не должен содержать пробелов");
-        notCorrectUser.setLogin("JavaGod");
-        notCorrectUser.setEmail("");
-        final ConstraintViolationException emailEmptyException = assertThrows(
-                ConstraintViolationException.class, () -> validateInput(notCorrectUser));
-        assertEquals(emailEmptyException.getMessage().substring(7), "Email не должен быть пустым");
-        notCorrectUser.setEmail("fgfhfgf");
-        final ConstraintViolationException notEmailException = assertThrows(
-                ConstraintViolationException.class, () -> validateInput(notCorrectUser));
-        assertEquals(notEmailException.getMessage().substring(7),
-                "Введённая строка не обладает структурой email [***@**.**]");
-        notCorrectUser.setEmail("java@yandex.ru");
-        notCorrectUser.setBirthday(LocalDate.MAX);
-        final ConstraintViolationException birthdayException = assertThrows(
-                ConstraintViolationException.class, () -> validateInput(notCorrectUser));
-        assertEquals(birthdayException.getMessage().substring(10),
-                "День рождения пользователя должен быть в прошлом.");
-        notCorrectUser.setBirthday(LocalDate.of(1990, 12, 12));
-        uc.setUser(notCorrectUser);
-        final BeanAlreadyCreatedException exception = assertThrows(
-                BeanAlreadyCreatedException.class, () -> uc.setUser(notCorrectUser));
-        assertEquals("User with id="+notCorrectUser.getId()+" is already added.", exception.getMessage());
-    }
-
-    @Test
-    void putUsersTest() {
-        UserController uc = new UserController();
-        User user = new User(1, "JavaGod", "Hacker","java@yandex.ru",
-                LocalDate.of(1990, 12, 12));
-        final BeanNotFoundException exception = assertThrows(
-                BeanNotFoundException.class,
-                () -> uc.putUser(user));
-        assertEquals("User with id="+user.getId()+" not found", exception.getMessage());
-        uc.setUser(user);
-        User user2 = new User(1, "PythonGod", "Junior","python@yandex.ru",
-                LocalDate.of(2000, 1, 1));
-        uc.putUser(user2);
-        assertEquals(uc.getAllUsers().get(0), user2);
-    }
-
     void validateInput(Film film) {
         ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
         Validator validator = factory.getValidator();
         Set<ConstraintViolation<Film>> violations = validator.validate(film);
-        if (!violations.isEmpty()) {
-            throw new ConstraintViolationException(violations);
-        }
-    }
-
-    void validateInput(User user) {
-        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
-        Validator validator = factory.getValidator();
-        Set<ConstraintViolation<User>> violations = validator.validate(user);
         if (!violations.isEmpty()) {
             throw new ConstraintViolationException(violations);
         }
